requirements.txt
Flask==1.1.1
Pillow==6.1.0
matplotlib==3.1.1 
Keras==2.3.0 
opencv-python==4.1.1.26 
vaderSentiment==3.2.1 
simplejson==3.16.0 